backend/requirements.txt
fastapi==0.109.0
mangum==0.17.0
pymupdf==1.23.21
google-generativeai>=0.7.2
python-multipart==0.0.7
Pillow==10.2.0